Initial commit
Welcome to homeR.  I created this package to productionalize some
R functions I built while selling and buying a home.  This is just the
basic function.  I have ideas to expand this function, expand to more than
just data gathering and expanding to more than just Zillow.  Not sure
where this will lead, but this is a start.

-Bill