API key to .Renviron file
Moved the API storage to the .Renviron startup file located in the users
HOME directory.  This is easier and more efficient than asking the user to create a random file in a random location.  The API key will be loaded everytime the user starts R.

The vignette was updated with instructions for creating and updating the
~/.Renviron file.

All remnants of the API key file have been removed.

Ready for version 1 to github.